environment_variableにValidateを追加する #5 change: EnvironmentVariableParseに渡された構造体にValidate関数が定義されている場合には実行するように変更 change: jsとwasip1でビルド対象とならないようにビルドタグを追加 test: テストコードをunix用・Windows用ではなく、ケースセンシティブ・ケースインセンシティブに変更
